// packages/core/src/adapters.generated.ts
export const GENERATED_ADAPTERS: Readonly<Record<string, DerivedAdapter>> = {
// ...
  "gmail": {
// ...
    "operations": {
      "gmail.users.drafts.create": {
        "request": {
          "kind": "http",
          "method": "POST",
          "pathTemplate": "/gmail/v1/users/{userId}/drafts",
          "pathParams": [
            "userId"
          ],
          "pathDefaults": {
            "userId": "me"
          },
          "queryParams": [],
          "hasBody": true
        },
        "scopes": [
          "https://mail.google.com/",
          "https://www.googleapis.com/auth/gmail.addons.current.action.compose",
          "https://www.googleapis.com/auth/gmail.compose",
          "https://www.googleapis.com/auth/gmail.modify"
        ]
      },
// ...
    }
  }
};